server_requirements.txt
fastapi
pydantic
uvicorn
PyVirtualDisplay
DrissionPage==4.0.5.6
websockets
loguru
slowapi
pyyaml